Autonomous Navigation and Mapping
* Most of the `perception_step` modifications were copies from the
Jupyter Notebook with 'data' replaced by 'Rover'.  A restriction was
added to worldmap updating so that the navigable, obstacle, and rock
pixels are only recorded on the map if the rover pitch and roll are
very small (less than 1 and 1.5deg respectively).  If the rover is
within view of a rock sample, Rover.mode was set to 'rock' to initiate
retrieval, and the nav_dists and nav_angles updates were calculated
using only the rock pixels rather than the navigable terrain pixels to
drive the rover directly towards the rock rather than continuing its
nominal path.

* In `decision_step`, the following modifications were made to force
the rover to follow the left wall (maze-solving technique): steering
angle was biased by taking the mean of only the 75% most left-aligned
nav_angles rather than the full set, 'stop' mode turned the rover to
the right to find the next left-handed wall, and 'reverse' mode steered
to the left when throttling backward and to the right when throttling
forward to keep obstacles to the rover's left when resuming forward
motion.  To avoid erratic drunk driving behavior, steering angles were
clipped to -10,10deg at max velocity.  Rover.mode = 'reverse' was added
to clear the rover from obstacles by attempting a series of extraction
maneuvers and exiting the mode when the rover achieves velocity > .5 or
runs out of moves; this mode is triggered when velocity = 0 for 5 or
more frames.  Rover.mode = 'rock' was added to steer the rover towards
any rock in view and execute the pickup; this mode is triggered by the
`perception_step`.

* The RoverState class in `drive_rover.py` was modified as follows:
- self.stuck added to determine whether the rover was stuck on an
obstacle by counting the number of consecutive velocity = 0 frames;
sets Rover.mode='reverse'
- self.reverse_ct added to track the number of cycles executed in the
obstacle extraction routine (Rover.mode='reverse')
- self.angle_ct added to count consecutive large left turn commands at
high speed (aka rover driving in a donut)
- stop_forward threshold was changed to 250 to keep the rover from
running into obstacles as frequently