Adjusted thresholds to improve rock pickup performance
In decision.py, reduced bias in the steering angle to minimize running
into walls, changed ‘rock’ mode behavior from ‘reverse’ to ‘stop’ when
the rover is stuck, and increased the reverse throttle to slow down
faster when the rover is moving too quickly towards the rock sample.

In drive_rover.py, adjusted stop_forward and go_forward thresholds to
be more lenient in order to avoid missing rocks.

In perception.py, added a rover in motion threshold to worldmap
recording to improve fidelity of the worldmap, and decreased the rock
sample pixel threshold for enabling ‘rock’ mode.